1.1.3
+ Added support to use decimal values for ranks & prestiges costs
+ Fixed error on displaying rankup

+++ src/main/java/me/drawethree/ultraprisoncore/ranks/rank/Prestige.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class Prestige extends Rank {
 
-    public Prestige(int id, long cost, String prefix, List<String> commandsToExecute) {
+    public Prestige(int id, double cost, String prefix, List<String> commandsToExecute) {
         super(id, cost, prefix, commandsToExecute);
     }
 }
